requirements.txt
pillow==11.1.0
pandas==2.2.3
openpyxl==3.1.5